requirements.txt
# Core dependencies
pdfplumber>=0.11.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Optional dependencies for enhanced functionality
openpyxl>=3.1.0  # For Excel output support
tabulate>=0.9.0  # For formatted table output

# Development dependencies (uncomment if needed)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
